refactor(kube): Resolver core를 internal/kube로 승격
informer/IP 인덱스/identity 팩토리/식별 병합 유틸을 internal/kube로 옮겨 netobs와 gpuobs가 동일한 IP→PodIdentity 인덱스를 공유할 수 있게 한다. netobs/metadata.Resolver는 *kube.Resolver를 embed해 IP 해석을 위임하고 그 위에 socket-cookie flow / cgroup·ifindex hint 캐시만 보유한다.
